+++ packages/testing/src/mod.ts
@@ -8,10 +8,7 @@
  * @module
  */
 
-export { MockContext, MockFederation } from "./mock.ts";
-export {
-  createContext,
-  createInboxContext,
-  createRequestContext,
-} from "./context.ts";
 export type { SentActivity } from "./mock.ts";
+export { MockContext, MockFederation } from "./mock.ts";
+export { mockDocumentLoader } from "./docloader.ts";
+export { createContext } from "./context.ts";